requirements.txt
beautifulsoup4==4.12.3
google==3.0.0
hashring==1.5.1
protobuf==3.20.1
python-memcached==1.62
six==1.16.0
soupsieve==2.5
tqdm==4.66.4
